feat: split localize out to its own function which is called in 'ready' hook
