+++ src/module/config/powers.ts
@@ -6,6 +6,11 @@ export const POWER_TYPES: CosmereAPI.PowerTypeConfigData[] = [
 export function register() {
 	POWER_TYPES.forEach(powerConfig => {
 		cosmereRPG.api.registerPowerType({ ...powerConfig, source: MODULE_ID });
+	});
+}
+
+export function localize() {
+	POWER_TYPES.forEach(powerConfig => {
 		// @ts-ignore
 		game.i18n!.translations.COSMERE.Powers[powerConfig.id] = powerConfig.label;
 	});
